memory optimization: Optimize CES memory usage by container improvements
- Replace std::map with std::unordered_map for O(1) lookup performance
- Replace std::set with std::vector to reduce memory overhead
- Compress deliveryState vector from int to uint8_t (4x space saving)
- Optimize GetTopSubscriberCounts algorithm from O(n^2) to O(n)
- Remove unused #include <list>

+++ services/include/ordered_event_record.h
@@ -44,7 +44,7 @@ struct OrderedEventRecord : public CommonEventRecord {
     int64_t finishTime;
     sptr<IRemoteObject> resultTo;
     sptr<IRemoteObject> curReceiver;
-    std::vector<int> deliveryState;
+    std::vector<uint8_t> deliveryState;
     std::vector<std::shared_ptr<EventSubscriberRecord>> receivers;
 
     OrderedEventRecord()
